diary 오류 다 같이 수정
