+++ src/main/java/com/example/spring_study/DTO/ClickDate.java
@@ -1,6 +1,5 @@
 package com.example.spring_study.DTO;
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
