+++ src/main/java/com/example/spring_study/Service/ScheduleService.java
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.time.LocalDate;
 import java.util.List;
-import com.example.spring_study.Dto.ScheduleDto;
+import com.example.spring_study.DTO.ScheduleDto;
 import com.example.spring_study.Repository.UserRepository;
 import org.springframework.transaction.annotation.Transactional;
